refactor: move the dataset to peer dependencies
This makes the process more flexible for consumers of the data and/or the API. Instead of having the
dataset installed at the specific version defined in the SDK, they can install the dataset at
whatever version they please. This way, the SDK doesn't need to be updated every time the dataset
gets updated.

BREAKING CHANGE: The SDK no longer includes the core dataset