refactor(getskills): always return an array
Previously, `getSkills` would return an object if `pokemonIDs` were provided. This made sense until
we added `pokemonID` to all of the skills in the core dataset. Now we'll return skills from
`getSkills` as an array with all of the matching skills.

BREAKING CHANGE: `getSkills` now has a different return signature.